+++ apps/charterafrica/payload.config.ts
@@ -9,6 +9,7 @@ import { buildConfig } from "payload/config";
 
 import Authors from "./src/payload/collections/Authors";
 import CommunityPlatforms from "./src/payload/collections/CommunityPlatforms";
+import CommonLabels from "./src/payload/globals/CommonLabels";
 import Events from "./src/payload/collections/Events";
 import Explainers from "./src/payload/collections/Explainers";
 import Fellowships from "./src/payload/collections/Fellowships";
@@ -91,6 +92,7 @@ export default buildConfig({
     Tools,
   ] as CollectionConfig[],
   globals: [
+    CommonLabels,
     FocalCountries,
     Footer,
     Helpdesk,
